Remove fixExportClass (#107)
* Added failing tests.

* I'm not entirely sure what fixExportClass.js was supposed to do, in any case the current version of TS handles export class correctly.
Some new ts helpers were added, so the new version caused some things to fail.
I've changed it to use the helpers in the offical tslib package. Also there is now an "importHelper" option which only
import the correct ones, so it isn't necessary to import them in the plugin. I've changed the helperId to "tslib", and am doing the
resolve myself so it doesn't fail if the user isn't using rollup-plugin-node-resolve. I'm not entirely sure this is the best way of
doing it.

* Update package.json

* Make it node 7 compatible